last

+++ MuscleTrainingRecords00/MuscleTrainingRecords00/GraphPage.xaml.cs
@@ -40,8 +40,8 @@ namespace MuscleTrainingRecords00
                
 
                 var db = TodoItemDatabase.getDatabase();
-                int B_Weight = int.Parse(bWeight.Text);
-                int B_Fat = int.Parse(bFat.Text);
+                double B_Weight = double.Parse(bWeight.Text);
+                double B_Fat = double.Parse(bFat.Text);
                 DateTime dCreated = DateTime.Today;
